+++ src/app/esri-map/esri-map.component.spec.ts
@@ -4,14 +4,28 @@ import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 
 import { EsriMapComponent } from './esri-map.component';
+import { ActivatedRoute, Data } from '@angular/router';
 
 describe('EsriMapComponent', () => {
   let component: EsriMapComponent;
   let fixture: ComponentFixture<EsriMapComponent>;
+  let MockMapClass;
 
   beforeEach(async(() => {
+    // provide a mock map class to the component via a mock route
+    MockMapClass = jasmine.createSpy('Map');
     TestBed.configureTestingModule({
-      declarations: [ EsriMapComponent ]
+      declarations: [ EsriMapComponent ],
+      providers: [{
+        provide: ActivatedRoute,
+        useValue: {
+          snapshot: {
+            data: {
+              esriModules: [MockMapClass]
+            }
+          }
+        }
+      }]
     })
     .compileComponents();
   }));
@@ -23,6 +37,11 @@ describe('EsriMapComponent', () => {
   });
 
   it('should create', () => {
-    expect(component).toBeTruthy();
+    // verify that the map was initialized with expected options
+    expect(MockMapClass.calls.mostRecent().args[1]).toEqual({
+      center: [-118, 34.5],
+      zoom: 8,
+      basemap: 'dark-gray'
+    });
   });
 });