+++ src/app/esri-map/esri-map.component.ts
@@ -1,50 +1,34 @@
-import { Component, OnInit, ElementRef } from '@angular/core';
-import { EsriLoaderService } from '../esri-loader.service';
+import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 
 @Component({
   selector: 'app-esri-map',
   templateUrl: './esri-map.component.html',
-  styleUrls: ['./esri-map.component.css'],
-  providers: [EsriLoaderService]
+  styleUrls: ['./esri-map.component.css']
 })
 export class EsriMapComponent implements OnInit {
+  @ViewChild('map') mapEl: ElementRef;
+
   map: any;
 
-  constructor(private elRef: ElementRef, private esriLoader: EsriLoaderService) {}
+  constructor(private route: ActivatedRoute) {}
 
   ngOnInit() {
     if (this.map) {
       // map is already initialized
       return;
     }
-    const esriLoaderOptions = {
-      // use a specific version of the API instead of the latest
-      url: '//js.arcgis.com/3.18/'
-    };
-    if (this.esriLoader.isLoaded()) {
-      this._createMap();
-    } else {
-      // must load ArcGIS API for JavaScript on the page before creating the map
-      this.esriLoader.load(esriLoaderOptions)
-      .then(() => {
-        this._createMap();
-      })
-      .catch(err => {
-        console.error(err);
-      });
-    }
+    // get the required esri classes from the route
+    const esriModules = this.route.snapshot.data['esriModules'];
+    this._createMap(esriModules);
   }
 
-  // load the map module and then
   // create a map at the root dom node of this component
-  _createMap() {
-    this.esriLoader.loadModules(['esri/map'])
-    .then((Map) => {
-      this.map = new Map(this.elRef.nativeElement.firstChild, {
-        center: [-118, 34.5],
-        zoom: 8,
-        basemap: 'dark-gray'
-      });
+  _createMap([Map]) {
+    this.map = new Map(this.mapEl.nativeElement, {
+      center: [-118, 34.5],
+      zoom: 8,
+      basemap: 'dark-gray'
     });
   }
 }