+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# angular-cli-esri Change Log
+# esri-angular-cli-example Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.2.0
 ## Changed
+- exracted EsriLoaderService into own repo angular2-esri-loader
 - extracted module loading code into esri-loader
 - EsriLoaderService functions return promises